Extract shared Roslyn worker assembly builder
Move worker compiler invocation, response-file construction, references,
quoting, cleanup, and .NET process policy out of the lifecycle host. Keep the
mutable test override in the host through a narrow seam adapter.
